+++ Carbon.IIS/Functions/Join-CIisVirtualPath.ps1
@@ -3,38 +3,51 @@ function Join-CIisVirtualPath
 {
     <#
     .SYNOPSIS
-    Combines a path and a child path for an IIS website, application, virtual directory into a single path.  
+    Combines a path and a child path for an IIS website, application, virtual directory into a single path.
 
     .DESCRIPTION
-    Removes extra slashes.  Converts backward slashes to forward slashes.  Relative portions are not removed.  Sorry.
-
-    Beginning with Carbon 2.0.1, this function is available only if IIS is installed.
+    Removes extra slashes and relative path signifiers (e.g. `.` and `..`).  Converts backward slashes to forward
+    slashes.
 
     .EXAMPLE
-    Join-CIisVirtualPath 'SiteName' 'Virtual/Path'
+    Join-CIisVirtualPath -Path 'SiteName' -ChildPath 'Virtual/Path'
 
-    Demonstrates how to join two IIS paths together.  REturns `SiteName/Virtual/Path`.
+    Demonstrates how to join two IIS paths together.  Returns `SiteName/Virtual/Path`.
     #>
     [CmdletBinding()]
     param(
-        [Parameter(Mandatory=$true,Position=0)]
-        [string]
         # The parent path.
-        $Path,
+        [Parameter(Mandatory, Position=0)]
+        [String]$Path,
 
-        [Parameter(Position=1)]
-        [string]
-        $ChildPath
+        # The child path.
+        [Parameter(Mandatory, Position=1)]
+        [AllowEmptyString()]
+        [AllowNull()]
+        [String] $ChildPath
     )
 
     Set-StrictMode -Version 'Latest'
-
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
+    $Path = $Path | ConvertTo-CIisVirtualPath -NoLeadingSlash
     if( $ChildPath )
     {
-        $Path = Join-Path -Path $Path -ChildPath $ChildPath
+        $ChildPath = $ChildPath | ConvertTo-CIisVirtualPath -NoLeadingSlash
     }
-    $Path.Replace('\', '/').Trim('/')
+
+    if( $Path -and $ChildPath )
+    {
+        return "$($Path)/$($ChildPath)"
+    }
+    elseif( $Path )
+    {
+        return $Path
+    }
+    elseif( $ChildPath )
+    {
+        return $ChildPath
+    }
+    return ''
 }