+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 
 If migrating from Carbon, follow these upgrade instructions.
 
+Upgrade to PowerShell 5.1 or PowerShell 7. PowerShell 4 is no longer supported. Microsoft's web administration API
+doesn't work under PowerShell 6, so neither does Carbon.IIS.
+
+Windows 2008 is no longer supported. Minimum operating system is Windows 8.1 and Windows Server 2012 R2.
+
 Replace usages of the `Install-CIisAppPool` function's `-UserName` and `-Password` arguments with the `-Credential`
 parameter:
 
@@ -29,15 +34,35 @@ usages to use the `GetAttributeValue` method instead, e.g. `GetAttributeValue('c
 `GetAttributeValue('destination')`, etc. The value of `httpResponseStatus` is now returned as an int, instead of an
 enumeration.
 
-Upgrade to PowerShell 5.1 or PowerShell 7. PowerShell 4 is no longer supported. Microsoft's web administration API
-doesn't work under PowerShell 6, so neither does Carbon.IIS.
+The `Install-CIisWebsite` function no longer sets the default anonymous authentication username on a website to
+nothing. New websites will now have the default anonymous authentication username of `IUSR`. To preserve the previous
+behavior, use the `Set-CIisAnonymousAuthentication`.
+
+`Get-CIisWebsite` now writes an error if a specific website doesn't exist. If a usage doesn't care if the website
+exists or not, add `-ErrorAction SilentlyContinue` or `-ErrorAction Ignore`.
+
+The `Join-CIisVirtualPath` function's `ChildPath` parameter is now mandatory. If you have usages without a `ChildPath`
+parameter, switch to using the new `ConvertTo-CIisVirtualPath`.
 
 ## Added
 
 * Carbon.IIS now supports
     * Windows PowerShell 5.1 (on .NET Framework 4.6.2 and later) and PowerShell 7
     * Windows 8 and 10, and Windows Server 2012R2, 2016, and 2019.
+* Function `Set-CIisAnonymousAuthentication` for configuring anonymous authentication.
+* Function `Remove-CIisConfigurationAttribute` for removing attributes from configuration sections.
+* Function `ConvertTo-CIisVirtualPath` for normalizing a virtual path (i.e. removing duplicate slashes, ensuring
+directory separators are `/`, etc.).
 
+## Changes
+
+* `Get-CIisWebsite` now writes an error if a specific website doesn't exist.
+* The `Join-CIisVirtualPath` function's `ChildPath` parameter is now required. Usages that don't have a `ChildPath`
+argument should switch to `ConvertTo-CIisVirtualPath`.
+
+## Fixed
+
+* `Install-CIisWebsite` wouldn't save changes when an existing website doesn't define its default application.
 
 ## Removed